move functionality from handlers.go to controller.go
