+++ internal/k8s/handlers.go
@@ -11,7 +11,6 @@ import (
 	"github.com/nginxinc/kubernetes-ingress/pkg/apis/dos/v1beta1"
 
 	"github.com/golang/glog"
-	"github.com/nginxinc/kubernetes-ingress/internal/configs"
 	"github.com/nginxinc/kubernetes-ingress/internal/k8s/secrets"
 	v1 "k8s.io/api/core/v1"
 	networking "k8s.io/api/networking/v1"
@@ -764,172 +763,6 @@ func createNamespaceHandlers(lbc *LoadBalancerController) cache.ResourceEventHan
 			}
 		},
 	}
-}
-
-func (lbc *LoadBalancerController) getVirtualServerByVirtualServerRoute(vsr *conf_v1.VirtualServerRoute) (*conf_v1.VirtualServer, bool) {
-	virtualServerID := vsr.Status.ReferencedBy
-
-	vs, exists := lbc.configuration.virtualServers[virtualServerID]
-	return vs, exists
-}
-
-func (lbc *LoadBalancerController) processVSWeightChangesWithoutReload(vsOld *conf_v1.VirtualServer, vsNew *conf_v1.VirtualServer) {
-	var weightUpdates []configs.WeightUpdate
-	var splitClientsIndex int
-	variableNamer := configs.NewVSVariableNamer(vsNew)
-
-	for i, routeNew := range vsNew.Spec.Routes {
-		routeOld := vsOld.Spec.Routes[i]
-		for j, matchNew := range routeNew.Matches {
-			matchOld := routeOld.Matches[j]
-			if len(matchNew.Splits) == 2 {
-				if matchNew.Splits[0].Weight != matchOld.Splits[0].Weight && matchNew.Splits[1].Weight != matchOld.Splits[1].Weight {
-					if matchNew.Splits[0].Weight+matchNew.Splits[1].Weight != 100 {
-						glog.V(3).Infof("VirtualServer %v has invalid weights", vsNew.Name)
-						return
-					}
-					weightUpdates = append(weightUpdates, configs.WeightUpdate{
-						Zone:  variableNamer.GetNameOfKeyvalZoneForSplitClientIndex(splitClientsIndex),
-						Key:   variableNamer.GetNameOfKeyvalKeyForSplitClientIndex(splitClientsIndex),
-						Value: variableNamer.GetNameOfKeyOfMapForWeights(splitClientsIndex, matchNew.Splits[0].Weight, matchNew.Splits[1].Weight),
-					})
-				}
-				splitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-			} else if len(matchNew.Splits) > 0 {
-				splitClientsIndex++
-			}
-		}
-		if len(routeNew.Splits) == 2 {
-			if routeNew.Splits[0].Weight != routeOld.Splits[0].Weight && routeNew.Splits[1].Weight != routeOld.Splits[1].Weight {
-				if routeNew.Splits[0].Weight+routeNew.Splits[1].Weight != 100 {
-					glog.V(3).Infof("VirtualServer %v has invalid weights", vsNew.Name)
-					return
-				}
-				weightUpdates = append(weightUpdates, configs.WeightUpdate{
-					Zone:  variableNamer.GetNameOfKeyvalZoneForSplitClientIndex(splitClientsIndex),
-					Key:   variableNamer.GetNameOfKeyvalKeyForSplitClientIndex(splitClientsIndex),
-					Value: variableNamer.GetNameOfKeyOfMapForWeights(splitClientsIndex, routeNew.Splits[0].Weight, routeNew.Splits[1].Weight),
-				})
-				splitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-			}
-			splitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-		} else if len(routeNew.Splits) > 0 {
-			splitClientsIndex++
-		}
-	}
-	for _, weight := range weightUpdates {
-		lbc.configurator.UpsertSplitClientsKeyVal(weight.Zone, weight.Key, weight.Value)
-	}
-}
-
-func (lbc *LoadBalancerController) processVSRWeightChangesWithoutReload(vsrOld *conf_v1.VirtualServerRoute, vsrNew *conf_v1.VirtualServerRoute) {
-	var weightUpdates []configs.WeightUpdate
-	vs, exists := lbc.getVirtualServerByVirtualServerRoute(vsrNew)
-	if !exists {
-		glog.V(3).Infof("VirtualServerRoute %v does not have a VirtualServer", vsrNew.Name)
-		return
-	}
-	splitClientsIndex := lbc.getStartingSplitClientsIndex(vsrNew, vs)
-
-	variableNamer := configs.NewVSVariableNamer(vs)
-
-	for i, routeNew := range vsrNew.Spec.Subroutes {
-		routeOld := vsrOld.Spec.Subroutes[i]
-		for j, matchNew := range routeNew.Matches {
-			matchOld := routeOld.Matches[j]
-			if len(matchNew.Splits) == 2 {
-				if matchNew.Splits[0].Weight != matchOld.Splits[0].Weight && matchNew.Splits[1].Weight != matchOld.Splits[1].Weight {
-					if matchNew.Splits[0].Weight+matchNew.Splits[1].Weight != 100 {
-						glog.V(3).Infof("VirtualServerRoute %v has invalid weights", vsrNew.Name)
-						return
-					}
-					weightUpdates = append(weightUpdates, configs.WeightUpdate{
-						Zone:  variableNamer.GetNameOfKeyvalZoneForSplitClientIndex(splitClientsIndex),
-						Key:   variableNamer.GetNameOfKeyvalKeyForSplitClientIndex(splitClientsIndex),
-						Value: variableNamer.GetNameOfKeyOfMapForWeights(splitClientsIndex, matchNew.Splits[0].Weight, matchNew.Splits[1].Weight),
-					})
-				}
-				splitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-			} else if len(matchNew.Splits) > 0 {
-				splitClientsIndex++
-			}
-		}
-		if len(routeNew.Splits) == 2 {
-			if routeNew.Splits[0].Weight != routeOld.Splits[0].Weight && routeNew.Splits[1].Weight != routeOld.Splits[1].Weight {
-				if routeNew.Splits[0].Weight+routeNew.Splits[1].Weight != 100 {
-					glog.V(3).Infof("VirtualServerRoute %v has invalid weights", vsrNew.Name)
-					return
-				}
-				weightUpdates = append(weightUpdates, configs.WeightUpdate{
-					Zone:  variableNamer.GetNameOfKeyvalZoneForSplitClientIndex(splitClientsIndex),
-					Key:   variableNamer.GetNameOfKeyvalKeyForSplitClientIndex(splitClientsIndex),
-					Value: variableNamer.GetNameOfKeyOfMapForWeights(splitClientsIndex, routeNew.Splits[0].Weight, routeNew.Splits[1].Weight),
-				})
-			}
-			splitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-		} else if len(routeNew.Splits) > 0 {
-			splitClientsIndex++
-		}
-	}
-	for _, weight := range weightUpdates {
-		lbc.configurator.UpsertSplitClientsKeyVal(weight.Zone, weight.Key, weight.Value)
-	}
-}
-
-func (lbc *LoadBalancerController) getVSExForVS(vs *conf_v1.VirtualServer) *configs.VirtualServerEx {
-	resources := lbc.configuration.GetResources()
-
-	resourceExes := lbc.createExtendedResources(resources)
-	for _, vsEx := range resourceExes.VirtualServerExes {
-		if vsEx.VirtualServer.Name == vs.Name {
-			return vsEx
-		}
-	}
-	return nil
-}
-
-func (lbc *LoadBalancerController) getStartingSplitClientsIndex(vsr *conf_v1.VirtualServerRoute, vs *conf_v1.VirtualServer) int {
-	var startingSplitClientsIndex int
-
-	for _, r := range vs.Spec.Routes {
-		for _, match := range r.Matches {
-			if len(match.Splits) == 2 {
-				startingSplitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-			} else if len(match.Splits) > 0 {
-				startingSplitClientsIndex++
-			}
-		}
-		if len(r.Splits) == 2 {
-			startingSplitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-		} else if len(r.Splits) > 0 {
-			startingSplitClientsIndex++
-		}
-
-	}
-
-	virtualserverEx := lbc.getVSExForVS(vs)
-
-	for _, vsRoute := range virtualserverEx.VirtualServerRoutes {
-		if vsRoute.Name == vsr.Name {
-			return startingSplitClientsIndex
-		}
-		for _, r := range vsRoute.Spec.Subroutes {
-			for _, match := range r.Matches {
-				if len(match.Splits) == 2 {
-					startingSplitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-				} else if len(match.Splits) > 0 {
-					startingSplitClientsIndex++
-				}
-			}
-			if len(r.Splits) == 2 {
-				startingSplitClientsIndex += splitClientAmountWhenWeightChangesWithoutReload
-			} else if len(r.Splits) > 0 {
-				startingSplitClientsIndex++
-			}
-		}
-	}
-
-	return startingSplitClientsIndex
 }
 
 func zeroOutVirtualServerSplitWeights(vs *conf_v1.VirtualServer) {
